It gets even cleaner if we make some custom datatypes. This allows us to: - set path only once instead of passing oit to every function again and again - cache results - define methods
Note that we defined structs, but also defined types as builtin datatypes, such as string (MemeFile).
This is helpfull if we want to
- document the true purpose of a variable (if it is a MemeFile that is far more descriptive than that it is a string)
- use type inference in our code
- define (custom) methods on these (custom) datatypes
  Note that this cannot be done on a list of cutom datatypes, but you can define a list of (custom) datatypes a a new custom datatype and set methods on that.
  as we do with custom datatype MemeFiles and its method getRandom().

Finaly, also note that we sometimes pass the object to its method byref instead of byvalue (e.a. *MemeDir instead of MemeDir).
If the method only reads and uses data, specify byvalue (e.a. MemeDir), if it modifies, or calls methods that do, specify byValue (*MemeDir)...